Getting started with syncing assault payload events

+++ cloak_and_dagger/Assets/Scripts/character/payload_carrier.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [RequireComponent(typeof(Collider2D), typeof(network_id))]
-public class payload_carrier : MonoBehaviour {
+public class payload_carrier : sync_behaviour<unit> {
 
 	[SerializeField]
 	private int_event_object payload_pickup;
@@ -12,12 +12,16 @@ public class payload_carrier : MonoBehaviour {
 	[SerializeField]
 	private int_event_object payload_delivered;
 
-	private payload payload_carried;
-	private network_id network_Id;
+	[SerializeField]
+	private int_event_object pre_local_death;
 
-	
-	void Start () {
-		network_Id = GetComponent<network_id>();
+	private payload payload_carried;
+
+
+	private void Awake() {
+		if (pre_local_death) {
+			pre_local_death.e.AddListener(on_pre_local_death);
+		}
 	}
 
 	private void OnTriggerEnter2D(Collider2D collision) {
@@ -29,23 +33,23 @@ public class payload_carrier : MonoBehaviour {
 			payload_carried.transform.SetParent(transform);
 			payload_carried.transform.localPosition = new Vector3(0.5f, 0);
 			payload_carried.transform.rotation = Quaternion.identity;
-			payload_pickup.Invoke(network_Id.val);
+			payload_pickup.Invoke(gameObject_id.val);
 		} else if (collision.gameObject.CompareTag("Payload_delivery_zone") && payload_carried != null) {
 			payload_delivery_zone PDZ = collision.gameObject.GetComponent<payload_delivery_zone>();
-			if (PDZ && PDZ.try_deliver_here((byte)network_Id.val)) {
+			if (PDZ && PDZ.try_deliver_here((byte)gameObject_id.val)) {
 				Destroy(payload_carried);
 				payload_carried = null;
-				payload_delivered.Invoke(network_Id.val);
+				payload_delivered.Invoke(gameObject_id.val);
 			}
 		}
 	}
 
-	private void OnDisable() {
-		if (payload_carried != null) {
+	private void on_pre_local_death(int player_id) {
+		if (is_local && payload_carried != null) {
 			payload_carried.transform.SetParent(null);
 			payload_carried.drop();
 			payload_carried = null;
-			payload_dropped.Invoke(network_Id.val);
+			payload_dropped.Invoke(gameObject_id.val);
 		}
 	}
